Refactor wallet and vendor services
